Add a TestCase class for minitest
