+++ lib/chatops/controller/test_case.rb
@@ -1,0 +1,5 @@
+require "chatops/controller/test_case_helpers"
+
+class ChatOps::Controller::TestCase < ActionController::TestCase
+  include ChatOps::Controller::TestCaseHelpers
+end
